add to error alert more messages and change ctms to eur

+++ src/auxiliar/I_Data_Access.java
@@ -7,9 +7,9 @@ import models.Currency;
 
 public interface I_Data_Access {
 	
-	public HashMap<Integer, Currency>  getCurrencyData();
+	public HashMap<Float, Currency>  getCurrencyData();
 	public HashMap<String, Catalog> getCatalogData();
-	public boolean saveCurrency(HashMap<Integer, Currency> currency);
+	public boolean saveCurrency(HashMap<Float, Currency> currency);
 	public boolean saveCatalog(HashMap<String, Catalog> catalog);
