+++ src/models/Currency.java
@@ -14,9 +14,9 @@ public class Currency {
 	 *  value - valor de la moneda
 	 *  amount - cantidad de montedas del mismo tipo
 	 */
-	private int id;
+	private float id;
 	private String  name; 
-	private int value;	
+	private float value;	
 	private int amount;
 	
 	
@@ -27,7 +27,7 @@ public class Currency {
 	 * @param value - valor de la moneda 
 	 * @param initialAmount - cantidad de monedas inicial
 	 */
-	public Currency(String name, int value, int initialAmount){
+	public Currency(String name, float value, int initialAmount){
 		this.id = value;
 		this.name = name;
 		this.value = value;
@@ -37,7 +37,7 @@ public class Currency {
 
 	// METODOS DE USO DE DATOS
 	
-	public int getTotalPrice() {
+	public float getTotalPrice() {
 		return value * amount;
 	}
 	
@@ -46,11 +46,11 @@ public class Currency {
 	/*
 	 * Metodos de recogida y el ajuste del id de la moneda
 	 */
-	public int getId() {
+	public float getId() {
 		return id;
 	}
 	
-	public void setId(int id) {
+	public void setId(float id) {
 		this.id = id;
 	}
 	
@@ -93,11 +93,11 @@ public class Currency {
 	 * Metodos de recogida y el ajuste de la cantidad de las monedas
 	 * 
 	 */
-	public void setValue(int value) {
+	public void setValue(float value) {
 		this.value = value;
 	}
 	
-	public int getValue() {
+	public float getValue() {
 		return value;
 	}
